+++ source/MaterialXGenArnold/ArnoldShaderGenerator.cpp
@@ -1,4 +1,4 @@
-#include <MaterialXGenOsl/ArnoldShaderGenerator.h>
+#include <MaterialXGenArnold/ArnoldShaderGenerator.h>
 
 namespace MaterialX
 {
@@ -6,8 +6,8 @@ namespace MaterialX
 const string ArnoldShaderGenerator::TARGET = "arnold";
 
 
-ArnoldShaderGenerator::ArnoldShaderGenerator() 
-    : OslShaderGenerator() 
+ArnoldShaderGenerator::ArnoldShaderGenerator()
+    : OslShaderGenerator()
 {
     const StringSet restrictedNames = { "translucent", "empirical_bssrdf", "randomwalk_bssrdf", "volume_absorption",
                                         "volume_emission", "volume_henyey_greenstein", "volume_matte" };
