+++ source/MaterialXGenArnold/ArnoldShaderGenerator.h
