+++ source/PyMaterialX/CMakeLists.txt
@@ -35,9 +35,13 @@ add_subdirectory(PyMaterialXGenShader)
 if (MATERIALX_BUILD_GEN_GLSL)
     add_subdirectory(PyMaterialXGenGlsl)
 endif()
+if (MATERIALX_BUILD_GEN_OSL)
+    add_subdirectory(PyMaterialXGenOsl)
+endif()
+
 if (MATERIALX_BUILD_GEN_OGSFX)
     add_subdirectory(PyMaterialXGenOgsFx)
 endif()
-if (MATERIALX_BUILD_GEN_OSL)
-    add_subdirectory(PyMaterialXGenOsl)
+if (MATERIALX_BUILD_GEN_ARNOLD)
+    add_subdirectory(PyMaterialXGenArnold)
 endif()
